Remove some code and reduce down complexity to single file
