+++ lib/minitest_world.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module MinitestWorld
+  def self.extended(base)
+    base.extend(Minitest::Assertions)
+    base.assertions = 0
+  end
+
+  attr_accessor :assertions
+end
