--- lib/multi_test/minitest_world.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-module MultiTest
-  module MinitestWorld
-    def self.extended(base)
-      base.extend(Minitest::Assertions)
-      base.assertions = 0
-    end
-
-    attr_accessor :assertions
-  end
-end
